refactor: check for config file existence

+++ app/update-core.sh
@@ -171,13 +171,20 @@ update_core_reset_plugins_js()
 	local CONFIG_PATH="$BRIDGECHAIN_PATH/packages/core/bin/config"
 	local PUBLISHED_CONFIG_PATH="$HOME/.config/$CHAIN_NAME-core"
 
-	mv "$PUBLISHED_CONFIG_PATH/mainnet/plugins.js" "$PUBLISHED_CONFIG_PATH/mainnet/plugins_2.3_$TIMESTAMP.js"
-	mv "$PUBLISHED_CONFIG_PATH/devnet/plugins.js" "$PUBLISHED_CONFIG_PATH/devnet/plugins_2.3_$TIMESTAMP.js"
-	mv "$PUBLISHED_CONFIG_PATH/testnet/plugins.js" "$PUBLISHED_CONFIG_PATH/testnet/plugins_2.3_$TIMESTAMP.js"
+	if [[ -f "$PUBLISHED_CONFIG_PATH/mainnet/plugins.js" ]]; then
+	    mv "$PUBLISHED_CONFIG_PATH/mainnet/plugins.js" "$PUBLISHED_CONFIG_PATH/mainnet/plugins_2.3_$timestamp.js"
+	    cp "$CONFIG_PATH/mainnet/plugins.js" "$PUBLISHED_CONFIG_PATH/mainnet/plugins.js"
+	fi
 
-	cp "$CONFIG_PATH/mainnet/plugins.js" "$PUBLISHED_CONFIG_PATH/mainnet/plugins.js"
-	cp "$CONFIG_PATH/devnet/plugins.js" "$PUBLISHED_CONFIG_PATH/devnet/plugins.js"
-	cp "$CONFIG_PATH/testnet/plugins.js" "$PUBLISHED_CONFIG_PATH/testnet/plugins.js"
+	if [[ -f "$PUBLISHED_CONFIG_PATH/devnet/plugins.js" ]]; then
+	    mv "$PUBLISHED_CONFIG_PATH/devnet/plugins.js" "$PUBLISHED_CONFIG_PATH/devnet/plugins_2.3_$timestamp.js"
+	    cp "$CONFIG_PATH/devnet/plugins.js" "$PUBLISHED_CONFIG_PATH/devnet/plugins.js"
+	fi
+
+	if [[ -f "$PUBLISHED_CONFIG_PATH/testnet/plugins.js" ]]; then
+	    mv "$PUBLISHED_CONFIG_PATH/testnet/plugins.js" "$PUBLISHED_CONFIG_PATH/testnet/plugins_2.3_$timestamp.js"
+	    cp "$CONFIG_PATH/testnet/plugins.js" "$PUBLISHED_CONFIG_PATH/testnet/plugins.js"
+	fi
 }
 
 update_core_prompt_to_push_changes()
